refactor: seed project domain package
