+++ internal/agentcompose/project/build.go
@@ -1,0 +1,194 @@
+package project
+
+import (
+	"encoding/json"
+	"fmt"
+	"strings"
+	"time"
+
+	cronlib "github.com/robfig/cron/v3"
+
+	"agent-compose/pkg/agentcompose/domain"
+	"agent-compose/pkg/compose"
+)
+
+const loaderDefaultCronTimezone = "UTC"
+
+type SchedulerTriggerBuild struct {
+	ID           string
+	Kind         string
+	Topic        string
+	IntervalMs   int64
+	SpecJSON     string
+	Registration string
+}
+
+type loaderCronSpec struct {
+	Kind     string `json:"kind,omitempty"`
+	Expr     string `json:"expr"`
+	Timezone string `json:"timezone,omitempty"`
+}
+
+var loaderCronParser = cronlib.NewParser(cronlib.SecondOptional | cronlib.Minute | cronlib.Hour | cronlib.Dom | cronlib.Month | cronlib.Dow | cronlib.Descriptor)
+
+func ManagedLoaderTriggersAndScript(projectID, agentName, schedulerName string, scheduler *compose.NormalizedSchedulerSpec) ([]SchedulerTriggerBuild, string, error) {
+	if scheduler == nil {
+		return nil, "", fmt.Errorf("scheduler is required")
+	}
+	triggers := make([]SchedulerTriggerBuild, 0, len(scheduler.Triggers))
+	seenNames := make(map[string]struct{}, len(scheduler.Triggers))
+	var script strings.Builder
+	script.WriteString("// Generated by agent-compose project scheduler reconcile.\n")
+	for i, trigger := range scheduler.Triggers {
+		name := strings.TrimSpace(trigger.Name)
+		if name != "" {
+			if _, ok := seenNames[name]; ok {
+				return nil, "", fmt.Errorf("duplicate scheduler trigger name %q", name)
+			}
+			seenNames[name] = struct{}{}
+		}
+		id, err := domain.StableManagedTriggerID(projectID, agentName, schedulerName, name, i)
+		if err != nil {
+			return nil, "", err
+		}
+		loaderTrigger, err := ManagedLoaderTriggerAndRegistration(id, agentName, trigger)
+		if err != nil {
+			return nil, "", err
+		}
+		triggers = append(triggers, loaderTrigger)
+		script.WriteString(loaderTrigger.Registration)
+	}
+	if len(triggers) == 0 {
+		script.WriteString("function main() { return { status: \"idle\" }; }\n")
+	}
+	return triggers, script.String(), nil
+}
+
+func ManagedLoaderTriggerAndRegistration(id, agentName string, trigger compose.NormalizedTriggerSpec) (SchedulerTriggerBuild, error) {
+	prompt := strings.TrimSpace(trigger.Prompt)
+	if prompt == "" {
+		prompt = fmt.Sprintf("Run agent %s.", agentName)
+	}
+	callback := fmt.Sprintf("async function(event) { return scheduler.agent(%s); }", JSStringLiteral(prompt))
+	switch trigger.Kind {
+	case "cron":
+		specJSON, err := loaderCronSpecJSON(trigger.Cron, "")
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		return SchedulerTriggerBuild{
+			ID:           id,
+			Kind:         domain.LoaderTriggerKindCron,
+			SpecJSON:     specJSON,
+			Registration: fmt.Sprintf("scheduler.cron(%s, %s, %s);\n", JSStringLiteral(id), JSStringLiteral(trigger.Cron), callback),
+		}, nil
+	case "interval":
+		interval, err := time.ParseDuration(trigger.Interval)
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		intervalMs := interval.Milliseconds()
+		if intervalMs <= 0 {
+			return SchedulerTriggerBuild{}, fmt.Errorf("interval trigger %s must be at least 1ms", id)
+		}
+		specJSON, err := marshalJSONCompact(map[string]any{"kind": domain.LoaderTriggerKindInterval, "interval": trigger.Interval})
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		return SchedulerTriggerBuild{
+			ID:           id,
+			Kind:         domain.LoaderTriggerKindInterval,
+			IntervalMs:   intervalMs,
+			SpecJSON:     specJSON,
+			Registration: fmt.Sprintf("scheduler.interval(%s, %s, %d);\n", JSStringLiteral(id), callback, intervalMs),
+		}, nil
+	case "timeout":
+		delay, err := time.ParseDuration(trigger.Timeout)
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		delayMs := delay.Milliseconds()
+		if delayMs <= 0 {
+			return SchedulerTriggerBuild{}, fmt.Errorf("timeout trigger %s must be at least 1ms", id)
+		}
+		specJSON, err := marshalJSONCompact(map[string]any{"kind": domain.LoaderTriggerKindTimeout, "timeout": trigger.Timeout})
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		return SchedulerTriggerBuild{
+			ID:           id,
+			Kind:         domain.LoaderTriggerKindTimeout,
+			IntervalMs:   delayMs,
+			SpecJSON:     specJSON,
+			Registration: fmt.Sprintf("scheduler.timeout(%s, %s, %d);\n", JSStringLiteral(id), callback, delayMs),
+		}, nil
+	case "event":
+		if trigger.Event == nil {
+			return SchedulerTriggerBuild{}, fmt.Errorf("event trigger topic is required")
+		}
+		topic := strings.TrimSpace(trigger.Event.Topic)
+		specJSON, err := marshalJSONCompact(map[string]any{"kind": domain.LoaderTriggerKindEvent, "topic": topic})
+		if err != nil {
+			return SchedulerTriggerBuild{}, err
+		}
+		return SchedulerTriggerBuild{
+			ID:           id,
+			Kind:         domain.LoaderTriggerKindEvent,
+			Topic:        topic,
+			SpecJSON:     specJSON,
+			Registration: fmt.Sprintf("scheduler.on(%s, %s, %s);\n", JSStringLiteral(topic), JSStringLiteral(id), callback),
+		}, nil
+	default:
+		return SchedulerTriggerBuild{}, fmt.Errorf("unsupported scheduler trigger kind %q", trigger.Kind)
+	}
+}
+
+func JSStringLiteral(value string) string {
+	data, err := json.Marshal(value)
+	if err != nil {
+		return `""`
+	}
+	return string(data)
+}
+
+func loaderCronSpecJSON(expr, timezone string) (string, error) {
+	spec, err := normalizeLoaderCronSpec(loaderCronSpec{
+		Kind:     domain.LoaderTriggerKindCron,
+		Expr:     expr,
+		Timezone: timezone,
+	})
+	if err != nil {
+		return "", err
+	}
+	return marshalJSONCompact(spec)
+}
+
+func normalizeLoaderCronSpec(spec loaderCronSpec) (loaderCronSpec, error) {
+	spec.Kind = domain.LoaderTriggerKindCron
+	spec.Expr = strings.TrimSpace(spec.Expr)
+	spec.Timezone = strings.TrimSpace(spec.Timezone)
+	if spec.Expr == "" {
+		return loaderCronSpec{}, fmt.Errorf("cron expr is required")
+	}
+	if spec.Timezone == "" {
+		spec.Timezone = loaderDefaultCronTimezone
+	}
+	if _, err := time.LoadLocation(spec.Timezone); err != nil {
+		return loaderCronSpec{}, fmt.Errorf("load cron timezone %q: %w", spec.Timezone, err)
+	}
+	if _, err := loaderCronParser.Parse(spec.Expr); err != nil {
+		return loaderCronSpec{}, fmt.Errorf("parse cron expression %q: %w", spec.Expr, err)
+	}
+	return spec, nil
+}
+
+func marshalJSONCompact(value any) (string, error) {
+	if value == nil {
+		return "", nil
+	}
+	data, err := json.Marshal(value)
+	if err != nil {
+		return "", fmt.Errorf("encode json payload: %w", err)
+	}
+	return string(data), nil
+}
