+++ pkg/agentcompose/project_build.go
@@ -1,12 +1,11 @@
 package agentcompose
 
 import (
-	"encoding/json"
 	"fmt"
 	"slices"
 	"strings"
-	"time"
 
+	projectdomain "agent-compose/internal/agentcompose/project"
 	"agent-compose/pkg/compose"
 )
 
@@ -163,121 +162,36 @@ func projectManagedLoaderFromScheduler(project ProjectRecord, scheduler ProjectS
 }
 
 func projectManagedLoaderTriggersAndScript(projectID, agentName, schedulerName string, scheduler *compose.NormalizedSchedulerSpec) ([]LoaderTrigger, string, error) {
-	if scheduler == nil {
-		return nil, "", fmt.Errorf("scheduler is required")
+	triggerBuilds, script, err := projectdomain.ManagedLoaderTriggersAndScript(projectID, agentName, schedulerName, scheduler)
+	if err != nil {
+		return nil, "", err
 	}
-	triggers := make([]LoaderTrigger, 0, len(scheduler.Triggers))
-	seenNames := make(map[string]struct{}, len(scheduler.Triggers))
-	var script strings.Builder
-	script.WriteString("// Generated by agent-compose project scheduler reconcile.\n")
-	for i, trigger := range scheduler.Triggers {
-		name := strings.TrimSpace(trigger.Name)
-		if name != "" {
-			if _, ok := seenNames[name]; ok {
-				return nil, "", fmt.Errorf("duplicate scheduler trigger name %q", name)
-			}
-			seenNames[name] = struct{}{}
-		}
-		id, err := StableManagedTriggerID(projectID, agentName, schedulerName, name, i)
-		if err != nil {
-			return nil, "", err
-		}
-		loaderTrigger, registration, err := projectManagedLoaderTriggerAndRegistration(id, agentName, trigger)
-		if err != nil {
-			return nil, "", err
-		}
-		triggers = append(triggers, loaderTrigger)
-		script.WriteString(registration)
+	triggers := make([]LoaderTrigger, 0, len(triggerBuilds))
+	for _, trigger := range triggerBuilds {
+		triggers = append(triggers, loaderTriggerFromProjectBuild(trigger))
 	}
-	if len(triggers) == 0 {
-		script.WriteString("function main() { return { status: \"idle\" }; }\n")
-	}
-	return triggers, script.String(), nil
+	return triggers, script, nil
 }
 
 func projectManagedLoaderTriggerAndRegistration(id, agentName string, trigger compose.NormalizedTriggerSpec) (LoaderTrigger, string, error) {
-	prompt := strings.TrimSpace(trigger.Prompt)
-	if prompt == "" {
-		prompt = fmt.Sprintf("Run agent %s.", agentName)
+	triggerBuild, err := projectdomain.ManagedLoaderTriggerAndRegistration(id, agentName, trigger)
+	if err != nil {
+		return LoaderTrigger{}, "", err
 	}
-	callback := fmt.Sprintf("async function(event) { return scheduler.agent(%s); }", jsStringLiteral(prompt))
-	switch trigger.Kind {
-	case "cron":
-		specJSON, err := loaderCronSpecJSON(trigger.Cron, "")
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		return LoaderTrigger{
-			ID:       id,
-			Kind:     LoaderTriggerKindCron,
-			Enabled:  true,
-			SpecJSON: specJSON,
-		}, fmt.Sprintf("scheduler.cron(%s, %s, %s);\n", jsStringLiteral(id), jsStringLiteral(trigger.Cron), callback), nil
-	case "interval":
-		interval, err := time.ParseDuration(trigger.Interval)
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		intervalMs := interval.Milliseconds()
-		if intervalMs <= 0 {
-			return LoaderTrigger{}, "", fmt.Errorf("interval trigger %s must be at least 1ms", id)
-		}
-		specJSON, err := marshalJSONCompact(map[string]any{"kind": LoaderTriggerKindInterval, "interval": trigger.Interval})
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		return LoaderTrigger{
-			ID:         id,
-			Kind:       LoaderTriggerKindInterval,
-			IntervalMs: intervalMs,
-			Enabled:    true,
-			SpecJSON:   specJSON,
-		}, fmt.Sprintf("scheduler.interval(%s, %s, %d);\n", jsStringLiteral(id), callback, intervalMs), nil
-	case "timeout":
-		delay, err := time.ParseDuration(trigger.Timeout)
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		delayMs := delay.Milliseconds()
-		if delayMs <= 0 {
-			return LoaderTrigger{}, "", fmt.Errorf("timeout trigger %s must be at least 1ms", id)
-		}
-		specJSON, err := marshalJSONCompact(map[string]any{"kind": LoaderTriggerKindTimeout, "timeout": trigger.Timeout})
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		return LoaderTrigger{
-			ID:         id,
-			Kind:       LoaderTriggerKindTimeout,
-			IntervalMs: delayMs,
-			Enabled:    true,
-			SpecJSON:   specJSON,
-		}, fmt.Sprintf("scheduler.timeout(%s, %s, %d);\n", jsStringLiteral(id), callback, delayMs), nil
-	case "event":
-		if trigger.Event == nil {
-			return LoaderTrigger{}, "", fmt.Errorf("event trigger topic is required")
-		}
-		topic := strings.TrimSpace(trigger.Event.Topic)
-		specJSON, err := marshalJSONCompact(map[string]any{"kind": LoaderTriggerKindEvent, "topic": topic})
-		if err != nil {
-			return LoaderTrigger{}, "", err
-		}
-		return LoaderTrigger{
-			ID:       id,
-			Kind:     LoaderTriggerKindEvent,
-			Topic:    topic,
-			Enabled:  true,
-			SpecJSON: specJSON,
-		}, fmt.Sprintf("scheduler.on(%s, %s, %s);\n", jsStringLiteral(topic), jsStringLiteral(id), callback), nil
-	default:
-		return LoaderTrigger{}, "", fmt.Errorf("unsupported scheduler trigger kind %q", trigger.Kind)
+	return loaderTriggerFromProjectBuild(triggerBuild), triggerBuild.Registration, nil
+}
+
+func loaderTriggerFromProjectBuild(trigger projectdomain.SchedulerTriggerBuild) LoaderTrigger {
+	return LoaderTrigger{
+		ID:         trigger.ID,
+		Kind:       trigger.Kind,
+		Topic:      trigger.Topic,
+		IntervalMs: trigger.IntervalMs,
+		Enabled:    true,
+		SpecJSON:   trigger.SpecJSON,
 	}
 }
 
 func jsStringLiteral(value string) string {
-	data, err := json.Marshal(value)
-	if err != nil {
-		return `""`
-	}
-	return string(data)
+	return projectdomain.JSStringLiteral(value)
 }
